chore: use maps to store ip and teid info
