--- internal/ue/gtp/service.go
@@ -1,104 +1,0 @@
-/**
- * SPDX-License-Identifier: Apache-2.0
- * © Copyright 2023 Hewlett Packard Enterprise Development LP
- */
-package gtp
-
-import (
-	"fmt"
-	"net"
-	"time"
-
-	"github.com/cilium/ebpf/link"
-	gnbContext "github.com/ellanetworks/core-tester/internal/gnb/context"
-	"github.com/ellanetworks/core-tester/internal/ue/context"
-	ebpf "github.com/ellanetworks/core-tester/internal/ue/ebpf"
-	log "github.com/sirupsen/logrus"
-)
-
-func SetupGtpInterface(ue *context.UEContext, msg gnbContext.UEMessage, n3Interface string, xdpAttachMode string) error {
-	gnbPduSession := msg.GNBPduSessions[0]
-	pduSession, err := ue.GetPduSession(uint8(gnbPduSession.GetPduSessionId()))
-	if err != nil {
-		return fmt.Errorf("failed to get PDU session: %w", err)
-	}
-	if pduSession == nil {
-		return fmt.Errorf("pdu session not found")
-	}
-
-	pduSession.GnbPduSession = gnbPduSession
-
-	if pduSession.Id != 1 {
-		return fmt.Errorf("pdu session id is not 1")
-	}
-
-	pduSession.SetGnbIp(msg.GnbIp)
-
-	ueGnbIp := pduSession.GetGnbIp()
-	upfIp := pduSession.GnbPduSession.GetUpfIp()
-	ueIp := pduSession.GetIp()
-	nameInf := "ellatester0"
-
-	time.Sleep(time.Second)
-
-	tunOpts := &TunnelOptions{
-		UEIP:             ueIp + "/16",
-		GTPUPort:         2152,
-		TunInterfaceName: nameInf,
-		GnbIP:            ueGnbIp.String(),
-		UpfIP:            upfIp,
-		Lteid:            gnbPduSession.GetTeidUplink(),
-		Rteid:            gnbPduSession.GetTeidDownlink(),
-	}
-	_, err = NewTunnel(tunOpts)
-	if err != nil {
-		return fmt.Errorf("failed to create tunnel: %w", err)
-	}
-
-	log.Infof("Created tunnel with options: %+v", tunOpts)
-
-	objs := ebpf.GTPObjects{}
-	if err := ebpf.LoadGTPObjects(&objs, nil); err != nil {
-		log.Fatalf("loading objects: %s", err)
-	}
-	defer objs.Close()
-
-	n3Iface, err := net.InterfaceByName(n3Interface)
-	if err != nil {
-		return fmt.Errorf("failed to lookup network iface %q: %s", n3Interface, err)
-	}
-
-	n3Link, err := link.AttachXDP(link.XDPOptions{
-		Program:   objs.Gtp,
-		Interface: n3Iface.Index,
-		Flags:     StringToXDPAttachMode(xdpAttachMode),
-	})
-	if err != nil {
-		return fmt.Errorf("failed to attach eBPF program on n3 interface %q: %s", n3Interface, err)
-	}
-	defer func() {
-		if err := n3Link.Close(); err != nil {
-			log.Warnf("Failed to detach eBPF program from n3 interface: %s", err)
-		}
-	}()
-
-	log.Infof("Attached eBPF program to n3 interface %q", n3Interface)
-
-	log.Info(fmt.Sprintf("[UE][GTP] Interface %s has successfully been configured for UE %s", nameInf, ueIp))
-	log.Info(fmt.Sprintf("[UE][GTP] You can do traffic for this UE by binding to IP %s, eg:", ueIp))
-	log.Info(fmt.Sprintf("[UE][GTP] iperf3 -B %s -c IPERF_SERVER -p PORT -t 9000", ueIp))
-	return nil
-}
-
-func StringToXDPAttachMode(Mode string) link.XDPAttachFlags {
-	switch Mode {
-	case "generic":
-		return link.XDPGenericMode
-	case "native":
-		return link.XDPDriverMode
-	case "offload":
-		return link.XDPOffloadMode
-	default:
-		return link.XDPGenericMode
-	}
-}
